Move context outside AppLayout
